requirements.txt
# Core Web Scraping
selenium>=4.33.0
webdriver-manager==4.0.1

# Google Sheets Integration
gspread>=6.2.1
oauth2client>=4.1.3
google-api-python-client>=2.127.0
google-auth>=2.29.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Additional Dependencies
pytz>=2024.1
requests>=2.31.0
# urllib3 will be installed automatically with compatible version

schedule==1.2.0
